Refactor gu_utils.py to tests/common
