+++ tests/generic_config_updater/conftest.py
@@ -3,8 +3,8 @@ import logging
 
 from tests.common.utilities import skip_release
 from tests.common.config_reload import config_reload
-from tests.generic_config_updater.gu_utils import apply_patch
-from tests.generic_config_updater.gu_utils import generate_tmpfile, delete_tmpfile
+from tests.common.gu_utils import apply_patch
+from tests.common.gu_utils import generate_tmpfile, delete_tmpfile
 
 CONFIG_DB = "/etc/sonic/config_db.json"
 CONFIG_DB_BACKUP = "/etc/sonic/config_db.json.before_gcu_test"
