+++ tests/generic_config_updater/test_bgpl.py
@@ -5,9 +5,9 @@ import re
 from netaddr import IPNetwork
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.helpers.generators import generate_ip_through_default_route
-from tests.generic_config_updater.gu_utils import apply_patch, expect_op_success
-from tests.generic_config_updater.gu_utils import generate_tmpfile, delete_tmpfile
-from tests.generic_config_updater.gu_utils import create_checkpoint, delete_checkpoint, rollback_or_reload
+from tests.common.gu_utils import apply_patch, expect_op_success
+from tests.common.gu_utils import generate_tmpfile, delete_tmpfile
+from tests.common.gu_utils import create_checkpoint, delete_checkpoint, rollback_or_reload
 
 pytestmark = [
     pytest.mark.topology('t0', 'm0', 'mx'),
