+++ tests/generic_config_updater/test_monitor_config.py
@@ -2,9 +2,9 @@ import logging
 import pytest
 
 from tests.common.helpers.assertions import pytest_assert
-from tests.generic_config_updater.gu_utils import apply_patch, expect_op_success, expect_res_success
-from tests.generic_config_updater.gu_utils import generate_tmpfile, delete_tmpfile
-from tests.generic_config_updater.gu_utils import create_checkpoint, delete_checkpoint, rollback, rollback_or_reload
+from tests.common.gu_utils import apply_patch, expect_op_success, expect_res_success
+from tests.common.gu_utils import generate_tmpfile, delete_tmpfile
+from tests.common.gu_utils import create_checkpoint, delete_checkpoint, rollback, rollback_or_reload
 
 pytestmark = [
     pytest.mark.topology('any'),
